refactor BoldedTitle component to a seperate file
